+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/entity/Trade.java
@@ -24,8 +24,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-import com.github.jnidzwetzki.bitfinex.v2.entity.symbol.BitfinexCurrencyPair;
-
 @Entity
 @Table(name = "trades")
 public class Trade {