+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/entity/symbol/BitfinexCandlestickSymbol.java
@@ -17,6 +17,7 @@
  *******************************************************************************/
 package com.github.jnidzwetzki.bitfinex.v2.entity.symbol;
 
+import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCurrencyPair;
 import com.github.jnidzwetzki.bitfinex.v2.entity.Timeframe;
 
 public class BitfinexCandlestickSymbol implements BitfinexStreamSymbol {