+++ src/test/java/com/github/jnidzwetzki/bitfinex/v2/test/OrderbookTest.java
@@ -22,10 +22,10 @@ import org.json.JSONTokener;
 import org.junit.Assert;
 import org.junit.Test;
 
+import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCurrencyPair;
 import com.github.jnidzwetzki.bitfinex.v2.entity.OrderBookFrequency;
 import com.github.jnidzwetzki.bitfinex.v2.entity.OrderBookPrecision;
 import com.github.jnidzwetzki.bitfinex.v2.entity.OrderbookConfiguration;
-import com.github.jnidzwetzki.bitfinex.v2.entity.symbol.BitfinexCurrencyPair;
 
 public class OrderbookTest {